repos/mvt_backend/requirements.txt
argcomplete==3.3.0
asgiref==3.7.2
boto3==1.34.125
botocore==1.34.125
certifi==2024.6.2
cfn-flip==1.3.0
charset-normalizer==3.3.2
click==8.1.7
Django==5.0.2
django-cors-headers==4.3.1
django-s3-storage==0.15.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
durationpy==0.6
exceptiongroup==1.2.2
hjson==3.1.0
hvac==2.2.0
idna==3.7
iniconfig==2.0.0
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.5
nose==1.3.7
packaging==24.1
placebo==0.9.0
pluggy==1.5.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
pytest==8.3.2
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
troposphere==4.8.0
typing_extensions==4.9.0
urllib3==2.2.1
Werkzeug==3.0.3
zappa==0.59.0
